avoid narrowing [unsigned] long long to [unsigned] long.  It's really useful on windows, where some types change  size when compiling to 64 bit

+++ include/fmt/printf.h
@@ -513,7 +513,7 @@ OutputIt basic_printf_context<OutputIt, Char>::format() {
         t = it != end ? *it : 0;
         convert_arg<long long>(arg, t);
       } else {
-        convert_arg<long>(arg, t);
+        convert_arg<long long>(arg, t);
       }
       break;
     case 'j':
